+++ modules/utils/isApp.js
@@ -1,5 +1,0 @@
-function isApp(object) {
-  return object && (typeof object === 'function' || typeof object.apply === 'function');
-}
-
-module.exports = isApp;