+++ CHANGES.md
@@ -3,6 +3,7 @@
   * Add onPart argument to multipart.Parser
   * Removed dependency on microtime module (issue #19)
   * Replace mach.urlMap with mach.map
+  * mach.file needs an app argument, same as other middleware
 
 ### 0.10.1 / 2014-07-06